requirements-dev.txt
#SYSTEM
fastapi==0.116.0
httpx==0.28.1
passlib==1.7.4
pydantic==2.11.7
pydantic_settings==2.10.1
pytest==8.4.1
pytest_asyncio==1.0.0
python-dotenv==1.1.1
python_jose==3.5.0
SQLAlchemy==2.0.41
uvicorn==0.30.6 
alembic==1.16.2
#TESTING
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
httpx==0.28.1
pytest-mock==3.10.0  